move Report module to a file structure, fix rejected promise warning
