+++ app/reports/channel_report.rb
@@ -1,0 +1,3 @@
+  class Report::Channel < Report::BaseReport
+  end
+
