+++ app/reports/invite_report.rb
@@ -1,0 +1,3 @@
+  class Report::Invite < Report::BaseReport
+  end
+
