+++ app/reports/organization_report.rb
@@ -1,0 +1,4 @@
+  class Report::Organization < Report::BaseReport
+  end
+
+
